+++ pkg/balancer/model.go
@@ -17,6 +17,41 @@
 
 package balancer
 
+//
+//                                           ,-------.
+//                                           |Initial|
+//                                      +----|-------|---+
+//                ProducerListenerReady |    `-------'   | ConsumerListenerReady
+//                                      v                v
+//                               ,-------------.   ,-------------.
+//                               |ProducerReady|   |ConsumerReady|
+//                               |-------------|   |-------------|
+//                               `-------------'   `-------------'
+//                ConsumerListenerReady |                | ProducerListenerReady
+//                                      v                v
+//                                   ,-----------------------.
+//                 LastProducerClose |NoConsumerAndNoProducer| LastConsumerClose
+// +-------------------------------->|-----------------------|<----------------------------------+
+// |                                 `-----------------------'                                   |
+// |                       NewConsumer |                 | NewProducer                           |
+// |                                   v                 v                                       |
+// |                          ,----------------.   ,----------------.                            |
+// |                          |ConsumerAwaiting|   |ProducerAwaiting|                            |
+// |                          |----------------|   |----------------|                            |
+// |                          `----------------'   `----------------'                            |
+// |          LastProducerClose |  NewPro. |           | NewCons. | LastConsumerClose            |
+// |                            v          v           v          v                              |
+// | ,-----------------------------.   ,-------------------.   ,-----------------------------.   |
+// | |ConsumerAwaitingAndCloseFirst|   |ConsumerAndProcuder|   |ProducerAwaitingAndCloseFirst|   |
+// | |-----------------------------|   |-------------------|   |-----------------------------|   |
+// | `-----------------------------'   `-------------------'   `-----------------------------'   |
+// |     NewProducer |                    |           |                        | NewConsumer     |
+// |                 v  LastConsumerClose |           | LastProducerClose      v                 |
+// |       ,------------------.           |           |              ,------------------.        |
+// |       |ConsumerCloseFirst|           |           |              |ProducerCloseFirst|        |
+// +-------|------------------|<----------+           +------------->|------------------|--------+
+//         `------------------'                                      `------------------'
+
 import (
 	"bufio"
 	"net"
@@ -51,6 +86,8 @@ const (
 	ConsumerAndProcuder
 	ConsumerCloseFirst
 	ProducerCloseFirst
+	ConsumerAwaitingAndCloseFirst
+	ProducerAwaitingAndCloseFirst
 )
 
 type StateEvent struct {
@@ -123,8 +160,19 @@ func update(
 		newState = ProducerCloseFirst
 	case StateEvent{state: ProducerCloseFirst, event: LastConsumerClose}:
 		newState = NoConsumerAndNoProducer
+
+	case StateEvent{state: ProducerAwaiting, event: LastProducerClose}:
+		newState = ProducerAwaitingAndCloseFirst
+	case StateEvent{state: ProducerAwaitingAndCloseFirst, event: NewConsumer}:
+		newState = ProducerCloseFirst
+
+	case StateEvent{state: ConsumerAwaiting, event: LastProducerClose}:
+		newState = ConsumerAwaitingAndCloseFirst
+	case StateEvent{state: ConsumerAwaitingAndCloseFirst, event: NewProducer}:
+		newState = ConsumerCloseFirst
+
 	default:
-		log.Logger.Debug().Int("state", int(state)).Int("event", int(event)).Msg("Unknow state and event action")
+		log.Logger.Warn().Int("state", int(state)).Int("event", int(event)).Msg("Unknow state and event action")
 
 		newState = state
 	}